ensure `UserDebugUtilsCallbackData.pMessageIdName` is not NULL
According to https://www.khronos.org/registry/vulkan/specs/1.2-extensions/man/html/VkDebugUtilsMessengerCallbackDataEXT.html:

```
pMessageIdName is a null-terminated string that identifies the particular message ID that is associated with the provided message. If the message corresponds to a validation layer message, then this string may contain the portion of the Vulkan specification that is believed to have been violated.
```

So NULL does not seem to be conformant. The new string uses a format similar to the validation layers.